requirements.txt
matplotlib==3.10.1
numpy==2.2.4
opencv_python==4.11.0.86
Pillow==11.2.1
scikit_learn==1.6.1
timm==1.0.15
torch==2.6.0
torchvision==0.21.0
